用户排名

+++ CloudBlog-Common/src/main/java/com/cloudblog/common/pojo/Vo/UserAchievementVo.java
@@ -15,7 +15,7 @@ public class UserAchievementVo {
 
     private Integer collectCount;
 
-    private Integer rank;
+    private Long rank;
 
     List<CreativeProcess> creativeProcessList;
